feat(wamn-587f.3): an int64 is a JSON string in every codec path
int64_is_json_string states the rule once. The conversion used to apply to a
revision alone, so a plain int64 reached the wire as a JSON number while the
binding called it a string. Now every int64 result field converts to a string,
and every IR input field decodes as JsonInt64. The CRUD path already did both,
and the revision flag selects no spelling anywhere.

codec_ir_rust_type reads the contract type through column_type_of and
delegates to codec_rust_type, so one map answers for both paths. The revision
parameter is gone from the codec result emitters.

The SQL bind does not change. A static SQL parameter keeps its i64 type, and
the decoded value reaches the bind site through the unwrap that the codec
already emitted.

WMS declares pallet_count as int32, under the owner ruling that a count is an
application integer. The query casts count(*) to integer, the guest carries an
i32, and the SQLx metadata is prepared again. No package now carries a
non-revision int64 on the wire, so Receiving and the Acme overlay regenerate
with no change at all.

Closes wamn-wpvg.

+++ apps/wamn_wms/query/inventory_aggregate.sql
@@ -8,7 +8,7 @@ SELECT
     pallet.location_id,
     pallet_quantity.status,
     sum(pallet_quantity.quantity) AS quantity,
-    count(*) AS pallet_count
+    count(*)::integer AS pallet_count
 FROM pallet_quantity AS pallet_quantity
 JOIN pallet AS pallet
     ON pallet.id = pallet_quantity.pallet_id
